Move _http_get_download to network.download
